+++ src/main/java/es/upm/grise/profundizacion/HandleDocuments/Document.java
@@ -14,8 +14,13 @@ public class Document {
 	private String body;
 	
 	public Document() throws NonRecoverableError {
-		this.documentId = DocumentIdProvider.getInstance().getDocumentId();
+		this.documentId = DocumentIdProvider.getInstance().getDocumentId(DocumentIdProvider.getInstance().getActualDocumentId());
 	}
+
+	public Document(int id) throws NonRecoverableError {
+		this.documentId = id;
+	}
+
 
 	public void setTemplate(String template) {
 		this.template = template;