+++ src/main/java/es/upm/grise/profundizacion/HandleDocuments/DocumentIdProvider.java
@@ -35,17 +35,17 @@ public class DocumentIdProvider {
 
 		else {
 
-			instance = new DocumentIdProvider();
+			String path = System.getenv(ENVIRON);
+			instance = new DocumentIdProvider(path, "com.mysql.cj.jdbc.Driver", "url", "user", "password");
 			return instance;
 
 		}	
 	}
 
 	// Create the connection to the database
-	private DocumentIdProvider() throws NonRecoverableError {
+	public DocumentIdProvider(String path, String DBDriver, String urlParam, String userParam, String passwordParam) throws NonRecoverableError {
 
-		// If ENVIRON does not exist, null is returned
-		String path = System.getenv(ENVIRON);
+
 		
 		if (path == null) {
 
@@ -75,14 +75,14 @@ public class DocumentIdProvider {
 			}
 
 			// Get the DB username and password
-			String url = propertiesInFile.getProperty("url");
-			String username = propertiesInFile.getProperty("username");
-			String password = propertiesInFile.getProperty("password");
+			String url = propertiesInFile.getProperty(urlParam);
+			String username = propertiesInFile.getProperty(userParam);
+			String password = propertiesInFile.getProperty(passwordParam);
 
 			// Load DB driver
 			try {
 
-				Class.forName("com.mysql.jdbc.Driver").newInstance();
+				Class.forName(DBDriver).newInstance();
 
 			} catch (InstantiationException e) {
 
@@ -172,7 +172,7 @@ public class DocumentIdProvider {
 	}
 
 	// Return the next valid objectID
-	public int getDocumentId() throws NonRecoverableError {
+	public int getDocumentId(int lastId) throws NonRecoverableError {
 
 		documentId++;
 
@@ -184,7 +184,7 @@ public class DocumentIdProvider {
 		try {
 
 			PreparedStatement preparedStatement = connection.prepareStatement(query);
-			preparedStatement.setInt(1, documentId);
+			preparedStatement.setInt(1, lastId++);
 			numUpdatedRows = preparedStatement.executeUpdate();
 
 		} catch (SQLException e) {
@@ -202,8 +202,12 @@ public class DocumentIdProvider {
 			throw new NonRecoverableError();
 
 		}
+		documentId = lastId++;
 
-		return documentId;
+		return lastId++;
 
+	}
+	public int getActualDocumentId() {
+		return this.documentId;
 	}
 }