+++ src/test/java/es/upm/grise/profundizacion/HandleDocuments/SmokeTest.java
@@ -13,12 +13,22 @@ public class SmokeTest {
 	@Test
 	public void formatTemplateCorrectly() throws NonRecoverableError, RecoverableError {
 		
-		Document d = new Document();
+		Document d = new Document(1623);
 		d.setTemplate("DECLARATION");
 		d.setTitle("A");
 		d.setAuthor("B");
 		d.setBody("C");
 		assertEquals("DOCUMENT ID: 1623\n\nTitle : A\nAuthor: B\n\nC", d.getFormattedDocument());
+
+	}
+
+	@Test
+	public void idIncrements() throws NonRecoverableError {
+
+		DocumentIdProvider t = DocumentIdProvider.getInstance();
+		int expectedId = 1643;
+
+		assertEquals(expectedId, t.getDocumentId(expectedId-1));
 
 	}